+++ hzlim/changelog.md
@@ -1,5 +1,9 @@
 # ai360 HZLIM Practice
 
+## Practice 02 Septermber 2024
+- added 3 hidden layers to `LogisticRegressionModel`
+- added more `hyperparameters` for training
+
 ## Practice 29 August 2024
 - established file structure and env for deep learning through `setup` module
 - `setup` module will detect platform and install relevant basic libraries in `torch`